test: add BowlingFrameTests
