+++ README.md
@@ -21,16 +21,15 @@
 * BowlingFrame
     * 볼링 게임의 프레임을 관리한다.
     * FrameScore와 마지막 프레임 유무를 관리한다.
+    * 남은 투구 기회 판단은 다음과 같다.
+            * 1-9 프레임 경우, 첫번째 투구가 스트라이크면 프레임 종료
+            * 1-9 프레임 경우, 두번째 투구가 끝나면 프레임 종료
+            * 10 프레임 경우, 두번째 투구까지 score 합이 10이나 20이 아니면 프레임 종료 
+            * 10 프레임 경우, 세번째 투구가 끝나면 프레임 종료s
 * FrameScore
     * 볼링 프레임의 점수를 관리한다.
     * Score 리스트를 관리한다.
     * Score 의 합을 반환한다.
-    * 남은 투구 기회가 있는지 반환한다.
-    * 남은 투구 기회 판단은 다음과 같다.
-        * 1-9 프레임 경우, 첫번째 투구가 스트라이크면 프레임 종료
-        * 1-9 프레임 경우, 두번째 투구가 끝나면 프레임 종료
-        * 10 프레임 경우, 두번째 투구까지 score 합이 10이나 20이 아니면 프레임 종료 
-        * 10 프레임 경우, 세번째 투구가 끝나면 프레임 종료
 * FrameScoreResult
     * FrameScore를 가지고 아래 4가지 종류를 표현한다.
     * 스트라이크(strike) : 프레임의 첫번째 투구에서 모든 핀(10개)을 쓰러트린 상태 (Ex. X)
@@ -38,8 +37,6 @@
     * 미스(miss) : 프레임의 두번재 투구에서도 모든 핀이 쓰러지지 않은 상태 (Ex. 7|-)
     * 거터(gutter) : 핀을 하나도 쓰러트리지 못한 상태. 거터는 "-"로 표시 (Ex. -|-)
 * Score
-    * Score는 0이상 10이하의 정수 타입이다.
-    * 0 ~ 10 사이의 정수를 입력 받아 첫번째 투구 점수를 생성할 수 있다.
-    * 0 ~ (10 - 첫번째 투구 점수) 사이의 정수를 입력 받아 두번째 투구 점수를 생성할 수 있다.
-    * 0 ~ 10 사이의 정수를 입력 받아 마지막 프레임의 두번째 투구 점수를 생성할 수 있다.
-    * 0 ~ 10 사이의 정수를 입력 받아 마지막 프레임의 세번째 투구 점수를 생성할 수 있다.
+    * Score는 0 이상 10 이하의 정수 타입이다.
+    * 0 ~ 10 사이의 정수를 입력 받아 투구 점수를 생성할 수 있다.
+    * 0 ~ (10 - 이전 투구 점수) 사이의 정수를 입력 받아 다음 투구 점수를 생성할 수 있다.
